+++ src/main/java/com/socialising/services/controller/UserController.java
@@ -1,13 +1,16 @@
 package com.socialising.services.controller;
 
+import com.fasterxml.jackson.databind.util.JSONPObject;
 import com.socialising.services.model.ChangePasswordRequest;
 import com.socialising.services.model.Image;
 import com.socialising.services.model.Post;
 import com.socialising.services.model.User;
 import com.socialising.services.service.UserService;
 import lombok.RequiredArgsConstructor;
+import org.json.JSONObject;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
+import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
 import org.springframework.messaging.handler.annotation.MessageMapping;
 import org.springframework.messaging.handler.annotation.Payload;
@@ -29,11 +32,19 @@ public class UserController {
 
     private final UserService userDetailsService;
 
+    private boolean checkTokenValidity(String token) {
+        if (!token.contains("Bearer ") || token.length() < 8) {
+            log.info("It is an invalid token. Pass the valid token!");
+            return false;
+        }
+        return true;
+    }
+
     // Add a User manually (Done by ADMIN only)
     @PostMapping("addUser")
     @PreAuthorize("hasAuthority('admin:create')")
-    public User addUser(@RequestBody User user) {
-        return this.userDetailsService.addUser(user);
+    public User addUser(@RequestBody User user, @RequestHeader("Authorization") String token) {
+        return this.userDetailsService.addUser(user, token);
     }
 
     // Get All the Users
@@ -46,6 +57,9 @@ public class UserController {
     // Get User details
     @GetMapping("getUserDetails")
     public User getUserDetails(@RequestHeader("Authorization") String token) {
+        if (!checkTokenValidity(token)) {
+            return null;
+        }
         return userDetailsService.getUserDetails(token);
     }
 
@@ -69,14 +83,17 @@ public class UserController {
         return this.userDetailsService.getUserByPhoneNumber(phoneNumber);
     }
 
-    // Update user details
-//    @PutMapping("updateDetails")
-//    public User updateUserDetails(@RequestBody User user, @RequestHeader("Authorization") String token) {
-//        return this.userDetailsService.updateUserDetails(token, user);
-//    }
+     // Update user details
+    @PatchMapping("updateUserDetails")
+    public User updateUserDetails(@RequestBody User user, @RequestHeader("Authorization") String token) {
+        if (!checkTokenValidity(token)) {
+            return null;
+        }
+        return this.userDetailsService.updateUserDetailsExceptUsernamePasswordAndDP(user, token);
+    }
 
     // DELETE User by ID
-    @DeleteMapping("deleteUser/{userid}")
+    @DeleteMapping("deleteUser/{userId}")
     @PreAuthorize("hasAuthority('admin:delete')")
     public int deleteUser(@PathVariable Long userId) {
         return this.userDetailsService.deleteUser(userId);
@@ -97,88 +114,145 @@ public class UserController {
     // To Send the Friend Request from User to User {toUsername}
     @PostMapping("sendFriendRequest/{username}")
     public String sendFriendRequest(@PathVariable("username") String toUsername, @RequestHeader("Authorization") String token) {
+        if (!checkTokenValidity(token)) {
+            return "Token Invalid";
+        }
         return this.userDetailsService.sendFriendRequest(toUsername, token);
     }
 
     // To Accept the Friend Request of User {userRequestId} to User {userid}
     @PostMapping("acceptFriendRequest/{username}")
     public String acceptFriendRequest(@PathVariable("username") String fromUsername, @RequestHeader("Authorization") String token) {
+        if (!checkTokenValidity(token)) {
+            return "Token Invalid";
+        }
         return this.userDetailsService.acceptFriendRequest(fromUsername, token);
     }
 
     // To Remove/Delete the Friend Request from User {userRequestId} to User {userid}
     @DeleteMapping("deleteFriendRequest/{friendRequestUsername}")
     public String deleteFriendRequest(@PathVariable("friendRequestUsername") String fromUsername, @RequestHeader("Authorization") String token) {
+        if (!checkTokenValidity(token)) {
+            return "Token invalid";
+        }
         return this.userDetailsService.deleteFriendRequest(fromUsername, token);
     }
 
     @GetMapping("getFriendRequestUsers")
     public String[] getFriendRequestUsers(@RequestHeader("Authorization") String token) {
+        if (!checkTokenValidity(token)) {
+            return null;
+        }
         return this.userDetailsService.getFriendRequestUsers(token);
+    }
+
+    @GetMapping("getFriendsRequested")
+    public String[] getFriendsRequestedByUser(@RequestHeader("Authorization") String token) {
+        if (!checkTokenValidity(token)) {
+            return null;
+        }
+        return this.userDetailsService.getFriendsRequested(token);
     }
 
     @GetMapping("getFriends")
     public String[] getFriendsOfUser(@RequestHeader("Authorization") String token) {
+        if (!checkTokenValidity(token)) {
+            return null;
+        }
         return this.userDetailsService.getFriendsOfUser(token);
     }
 
     @DeleteMapping("deleteFriend/{friendUsername}")
     public int deleteFriend(@PathVariable("friendUsername") String friendUsername, @RequestHeader("Authorization") String token) {
+        if (!checkTokenValidity(token)) {
+            return -1;
+        }
         return this.userDetailsService.deleteFriend(friendUsername, token);
     }
 
     @GetMapping("getPosts")
     public List<Post> getPostsOfUser(@RequestHeader("Authorization") String token) {
+        if (!checkTokenValidity(token)) {
+            return null;
+        }
         return this.userDetailsService.getPostsOfUser(token);
     }
 
     @GetMapping("getRequestedPosts")
     public List<Post> getRequestedPostsOfUser(@RequestHeader("Authorization") String token) {
+        if (!checkTokenValidity(token)) {
+            return null;
+        }
         return this.userDetailsService.getRequestedPostsOfUser(token);
     }
 
     @GetMapping("getReminderPosts")
     public List<Post> getReminderPosts(@RequestHeader("Authorization") String token) {
+        if (!checkTokenValidity(token)) {
+            return null;
+        }
         return this.userDetailsService.getReminderPosts(token);
     }
 
     @DeleteMapping("deleteReminderPost/{postId}")
     public List<Post> deleteReminderPostsOfUser(@PathVariable("postId") Long postId, @RequestHeader("Authorization") String token) {
+        if (!checkTokenValidity(token)) {
+            return null;
+        }
         return this.userDetailsService.deleteReminderPostsOfUser(postId, token);
     }
 
     @GetMapping("getTagsofUser")
     public String[] getTagsOfUser(@RequestHeader("Authorization") String token) {
+        if (!checkTokenValidity(token)) {
+            return null;
+        }
         return this.userDetailsService.getTagsOfUser(token);
     }
 
     @PutMapping("updateTagsOfUser")
     public String[] updateTagsOfUser(@RequestBody String[] newTags, @RequestHeader("Authorization") String token) {
+        if (!checkTokenValidity(token)) {
+            return null;
+        }
         return this.userDetailsService.updateTagsOfUser(newTags, token);
     }
 
     @PostMapping("addUserDP")
     public Image addUserDP(@RequestBody MultipartFile file, @RequestHeader("Authorization") String token) throws Exception {
+        if (!checkTokenValidity(token)) {
+            return null;
+        }
         return this.userDetailsService.addUserDP(file, token);
     }
 
     @GetMapping("getUserDP")
     public Image getImage(@RequestHeader("Authorization") String token) throws Exception {
+        if (!checkTokenValidity(token)) {
+            return null;
+        }
         return this.userDetailsService.getUserDP(token);
     }
 
     @DeleteMapping("removeDP")
     public int removeUserDP(@RequestHeader("Authorization") String token) {
+        if (!checkTokenValidity(token)) {
+            return -1;
+        }
         return this.userDetailsService.removeUserDP(token);
     }
 
     // CHAT based APIs
     @MessageMapping("/user.addUser")
     @SendTo("/user/public")
-    public User connectUser(@Payload User user) {
-        return this.userDetailsService.addUser(user);
+    public User connectUser(@Payload User user, @RequestHeader("Authorization") String token) {
+        if (!checkTokenValidity(token)) {
+            return null;
+        }
+        return this.userDetailsService.addUser(user, token);
     }
 
+    // All the users subscribed to /user/public will get to know that the user has disconnected
     @MessageMapping("/user.disconnectUser")
     @SendTo("/user/public")
     public User disconnect(@Payload User user) {
